Added wall tex in building map generation
